+++ x2gbfs/x2gbfs.py
@@ -28,6 +28,15 @@ logging.basicConfig()
 logging.getLogger().setLevel(logging.INFO)
 logger = logging.getLogger('x2gbfs')
 
+X2GBFS_DEFAULT_CONFIG = {
+    # Default ttl for every gbfs file.
+    # Might be overriden for static feeds that are expected to change less frequently.
+    'ttl': 60,
+    # Is this feed for protected publication? If true, param `protectedBaseUrl`
+    # must be provided on startup and will be used as feed base url instead of `baseUrl`
+    'protected': False,
+}
+
 
 def build_extractor(provider: str, feed_config: Dict[str, Any]) -> BaseProvider:
     if provider == 'example':
@@ -72,16 +81,20 @@ def build_extractor(provider: str, feed_config: Dict[str, Any]) -> BaseProvider:
     raise ValueError(f'Unknown config {provider}')
 
 
-def main(providers: List[str], output_dir: str, base_url: str, interval: int = 0) -> None:
+def main(
+    providers: List[str], output_dir: str, base_url: str, protected_base_url: str | None, interval: int = 0
+) -> None:
     should_loop_infinetly = interval > 0
     error_occured = False
 
     while True:
         for provider in providers:
             try:
-                generate_feed_for(provider, output_dir, base_url)
+                generate_feed_for(provider, output_dir, base_url, protected_base_url)
             except HTTPError as err:
-                logger.error(f'Generating feed for {provider} failed due to HTTP error {err.response.text}')
+                logger.error(
+                    f'Generating feed for {provider} failed due to HTTP error {err.response.status_code} for url {err.request.url}'
+                )
                 error_occured = True
             except TimeoutError:
                 logger.error(f'Generating feed for {provider} failed due to timeout error!')
@@ -97,7 +110,15 @@ def main(providers: List[str], output_dir: str, base_url: str, interval: int = 0
             exit(error_occured)
 
 
-def generate_feed_for(provider: str, output_dir: str, base_url: str) -> None:
+def get_x2gbfs_config_value(feed_config, key):
+    """
+    Returns x2gbfs config value, which can be overritten per feed via a x2gbfs section
+    in feed config.
+    """
+    return feed_config.get('x2gbfs', {}).get(key, X2GBFS_DEFAULT_CONFIG.get(key))
+
+
+def generate_feed_for(provider: str, output_dir: str, base_url: str, protected_base_url: str | None) -> None:
     with open(f'config/{provider}.json') as config_file:
         feed_config = json.load(config_file)
 
@@ -107,6 +128,17 @@ def generate_feed_for(provider: str, output_dir: str, base_url: str) -> None:
     (info, status, vehicle_types, vehicles, geofencing_zones, last_reported) = transformer.load_stations_and_vehicles(
         extractor
     )
+
+    is_feed_protected = get_x2gbfs_config_value(feed_config, 'protected')
+    if is_feed_protected:
+        if protected_base_url is None:
+            logger.warning(f'Feed {provider} is declared protected, but protectedBaseUrl is undefined, using baseUrl')
+            feed_base_url = f'{base_url}/{provider}'
+        else:
+            feed_base_url = f'{protected_base_url}/{provider}'
+    else:
+        feed_base_url = f'{base_url}/{provider}'
+
     GbfsWriter().write_gbfs_feed(
         feed_config,
         f'{output_dir}/{provider}',
@@ -115,9 +147,9 @@ def generate_feed_for(provider: str, output_dir: str, base_url: str) -> None:
         vehicle_types,
         vehicles,
         geofencing_zones,
-        f'{base_url}/{provider}',
+        feed_base_url,
         last_reported,
-        ttl=feed_config.get('ttl', 60),
+        ttl=get_x2gbfs_config_value(feed_config, 'ttl'),
     )
     logger.info(f'Updated feeds for {provider}')
 
@@ -128,7 +160,18 @@ if __name__ == '__main__':
         '-o', '--outputDir', help='output directory the transformed files are written to', default='out'
     )
     parser.add_argument('-p', '--providers', required=True, help='service provider(s), comma-separated')
-    parser.add_argument('-b', '--baseUrl', required=True, help='baseUrl this/these feed(s) will be published under')
+    parser.add_argument(
+        '-b',
+        '--baseUrl',
+        required=True,
+        help='baseUrl this/these feed(s) will be published under (if not flagged protected in config)',
+    )
+    parser.add_argument(
+        '-r',
+        '--protectedBaseUrl',
+        required=False,
+        help='protected baseUrl this/these feed(s) will be published under, if config declares `x2gbfs/protected: true`',
+    )
     parser.add_argument(
         '-i',
         '--interval',
@@ -140,4 +183,4 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
 
-    main(args.providers.split(','), args.outputDir, args.baseUrl, args.interval)
+    main(args.providers.split(','), args.outputDir, args.baseUrl, args.protectedBaseUrl, args.interval)